Remove reqwest blocking fix todo.
I messed with it for about an hour and couldn't make it work, because
reqwest was complaining about not running under the Tokio
runtime. Digging a bit, I'm going to chalk this up to different
versions - I'm using Tokio 0.3 and it is expecting 0.2.5. I did a
quick test to drop back to 0.2, and a bunch of other stuff broke.

Realistically, it's diminishing returns - it's already a set of
parallel tasks which are sequential. So, the fact that they block is
largely inconsequential.