Replace all instances of `expect`, `unwrap`, and `panic`.
The file handling semantics likely could be more memory efficient, but
reading the entire file into a String and then vectorizing it nicely
isolates the read code (which can generate I/O errors) from the
vectorization code (which would never fail). This isn't an embedded
system, we have a lot of RAM, and the files are small. Should be fine.